+++ src/ui/particle.rs
@@ -121,15 +121,15 @@ impl Builder {
 
         // if we have multiple particles, then render one per frame
         if self.end_pos.is_some() || self.end_col.is_some() {
-            let pos_start = rltk::PointF::new(self.pos.x as f32, self.pos.y as f32);
+            let pos_start = rltk::PointF::new(self.pos.x, self.pos.y);
 
             let mut t = 0.0;
             while t < self.lifetime {
                 let progress = t / self.lifetime;
                 let pos = self.end_pos.map_or(pos_start, |pos_end| {
                     rltk::PointF::new(
-                        progress.mul_add(pos_end.x as f32 - pos_start.x, pos_start.x),
-                        progress.mul_add(pos_end.y as f32 - pos_start.y, pos_start.y),
+                        progress.mul_add(pos_end.x - pos_start.x, pos_start.x),
+                        progress.mul_add(pos_end.y - pos_start.y, pos_start.y),
                     )
                 });
                 let col = self.end_col.map_or((self.col_fg, self.col_bg), |c| {
